✨[상품주문]: 상품 검색 모듈 분리
